+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Search now shows the paragraph where the searched term (highlighted) was found in the text.
+
+
 ## [0.1.0] - 2025-03-17
 - Defines the basic structure of the site with `index`, `single`, `page`, `taxonomy-series` and `search`.
 - Defines `parts` folder for storing template parts (i.e. header and footer) and `patterns` for storing template patterns like loops